Reconcile the discover freshness rule and the client's length with the branches beneath
The stateless stdio branch's sixth round settled how a DiscoverResult's
ttlMs is read (an absent, zero, negative or malformed hint is stale at
once) and that a stale discovery is refreshed before a capability is
judged; this branch's sixth round made the capability gate shared and the
clock the transport's own. One rule now serves both: the hint is read by
CachedResult.normalize_ttl for the discover entry as for every other, the
shared gate refreshes before it judges, and the round-39 pin follows the
no-hint reading. The client's cache-slice bookkeeping moves into
Client::CacheSlices to keep the class within its length budget after the
lower branches' additions.
